requirements.txt
gradio # Remove soon.
python-dotenv
aiohttp
openai
ijson
nest_asyncio
tenacity
azure-cosmos
python-decouple
streamlit
streamlit-chat
